Replace emitters
