+++ src/compiler/core/src/commonTest/kotlin/community/flock/wirespec/compiler/core/CompileEnumTest.kt
@@ -6,9 +6,11 @@ import community.flock.wirespec.compiler.core.emit.ScalaEmitter
 import community.flock.wirespec.compiler.core.emit.TypeScriptEmitter
 import community.flock.wirespec.compiler.core.emit.WirespecEmitter
 import community.flock.wirespec.compiler.core.emit.common.AbstractEmitter
+import community.flock.wirespec.compiler.core.emit.common.Emitter
 import community.flock.wirespec.compiler.utils.noLogger
 import io.kotest.assertions.arrow.core.shouldBeRight
 import kotlin.test.Test
+import kotlin.test.assertEquals
 
 class CompileEnumTest {
 
@@ -25,21 +27,22 @@ class CompileEnumTest {
     @Test
     fun testEnumKotlin() {
         val kotlin = """
-            package community.flock.wirespec.generated
-            
-            import community.flock.wirespec.Wirespec
-            
-            enum class MyAwesomeEnum (val label: String): Wirespec.Enum {
-              ONE("ONE"),
-              Two("Two"),
-              THREE_MORE("THREE_MORE");
-            
-              override fun toString(): String {
-                return label
-              }
-            }
-            
-        """.trimIndent()
+            |package community.flock.wirespec.generated
+            |
+            |import community.flock.wirespec.Wirespec
+            |
+            |enum class MyAwesomeEnum (val label: String): Wirespec.Enum {
+            |  ONE("ONE"),
+            |  Two("Two"),
+            |  THREE_MORE("THREE_MORE");
+            |  override fun toString(): String {
+            |    return label
+            |  }
+            |}
+            |
+        """.trimMargin()
+
+        assertEquals(kotlin, compiler(KotlinEmitter(logger = logger)).getOrNull())
 
         compiler(KotlinEmitter(logger = logger)) shouldBeRight kotlin
     }
@@ -47,25 +50,27 @@ class CompileEnumTest {
     @Test
     fun testEnumJava() {
         val java = """
-            package community.flock.wirespec.generated;
-            
-            import community.flock.wirespec.Wirespec;
-            
-            public enum MyAwesomeEnum implements Wirespec.Enum {
-              ONE("ONE"),
-              Two("Two"),
-              THREE_MORE("THREE_MORE");
-              public final String label;
-              MyAwesomeEnum(String label) {
-                this.label = label;
-              }
-              @Override
-              public String toString() {
-                return label;
-              }
-            }
+            |package community.flock.wirespec.generated;
+            |
+            |import community.flock.wirespec.Wirespec;
+            |
+            |public enum MyAwesomeEnum implements Wirespec.Enum {
+            |  ONE("ONE"),
+            |  Two("Two"),
+            |  THREE_MORE("THREE_MORE");
+            |  public final String label;
+            |  MyAwesomeEnum(String label) {
+            |    this.label = label;
+            |  }
+            |  @Override
+            |  public String toString() {
+            |    return label;
+            |  }
+            |}
+            |
+        """.trimMargin()
 
-        """.trimIndent()
+        assertEquals(java, compiler(JavaEmitter(logger = logger)).getOrNull())
 
         compiler(JavaEmitter(logger = logger)) shouldBeRight java
     }
@@ -109,7 +114,7 @@ class CompileEnumTest {
         compiler(WirespecEmitter(logger = logger)) shouldBeRight wirespec
     }
 
-    private fun compile(source: String) = { emitter: AbstractEmitter ->
+    private fun compile(source: String) = { emitter: Emitter ->
         Wirespec.compile(source)(logger)(emitter)
             .map { it.first().result }
             .onLeft(::println)
