+++ src/compiler/core/src/commonTest/kotlin/community/flock/wirespec/compiler/core/CompileRefinedTest.kt
@@ -5,7 +5,7 @@ import community.flock.wirespec.compiler.core.emit.KotlinEmitter
 import community.flock.wirespec.compiler.core.emit.ScalaEmitter
 import community.flock.wirespec.compiler.core.emit.TypeScriptEmitter
 import community.flock.wirespec.compiler.core.emit.WirespecEmitter
-import community.flock.wirespec.compiler.core.emit.common.AbstractEmitter
+import community.flock.wirespec.compiler.core.emit.common.Emitter
 import community.flock.wirespec.compiler.utils.noLogger
 import io.kotest.assertions.arrow.core.shouldBeRight
 import kotlin.test.Test
@@ -97,7 +97,7 @@ class CompileRefinedTest {
         compiler(WirespecEmitter(logger = logger)) shouldBeRight wirespec
     }
 
-    private fun compile(source: String) = { emitter: AbstractEmitter ->
+    private fun compile(source: String) = { emitter: Emitter ->
         Wirespec.compile(source)(logger)(emitter)
             .map { it.first().result }
             .onLeft(::println)
